requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.11.0
Flask-Migrate==4.0.7
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0